AP_InertialSensor: convert bandwidth and frequency to floats.

+++ libraries/AP_InertialSensor/AP_InertialSensor_Backend.h
@@ -239,10 +239,10 @@ protected:
     uint16_t get_sample_rate_hz(void) const;
 
     // return the notch filter center in Hz for the sample rate
-    uint16_t _gyro_notch_center_freq_hz(void) const { return _imu._notch_filter.center_freq_hz(); }
+    float _gyro_notch_center_freq_hz(void) const { return _imu._notch_filter.center_freq_hz(); }
 
     // return the notch filter bandwidth in Hz for the sample rate
-    uint16_t _gyro_notch_bandwidth_hz(void) const { return _imu._notch_filter.bandwidth_hz(); }
+    float _gyro_notch_bandwidth_hz(void) const { return _imu._notch_filter.bandwidth_hz(); }
 
     // return the notch filter attenuation in dB for the sample rate
     float _gyro_notch_attenuation_dB(void) const { return _imu._notch_filter.attenuation_dB(); }
@@ -250,10 +250,10 @@ protected:
     uint8_t _gyro_notch_enabled(void) const { return _imu._notch_filter.enabled(); }
 
     // return the dynamic notch filter center in Hz for the sample rate
-    uint16_t _gyro_harmonic_notch_center_freq_hz(uint8_t instance) const { return _imu._calculated_harmonic_notch_freq_hz[instance]; }
+    float _gyro_harmonic_notch_center_freq_hz(uint8_t instance) const { return _imu._calculated_harmonic_notch_freq_hz[instance]; }
 
     // return the notch filter bandwidth in Hz for the sample rate
-    uint16_t _gyro_harmonic_notch_bandwidth_hz(void) const { return _imu._harmonic_notch_filter.bandwidth_hz(); }
+    float _gyro_harmonic_notch_bandwidth_hz(void) const { return _imu._harmonic_notch_filter.bandwidth_hz(); }
 
     // return the notch filter attenuation in dB for the sample rate
     float _gyro_harmonic_notch_attenuation_dB(void) const { return _imu._harmonic_notch_filter.attenuation_dB(); }
@@ -269,13 +269,13 @@ protected:
     // support for updating filter at runtime
     uint16_t _last_accel_filter_hz[INS_MAX_INSTANCES];
     uint16_t _last_gyro_filter_hz[INS_MAX_INSTANCES];
-    uint16_t _last_notch_center_freq_hz[INS_MAX_INSTANCES];
-    uint16_t _last_notch_bandwidth_hz[INS_MAX_INSTANCES];
+    float _last_notch_center_freq_hz[INS_MAX_INSTANCES];
+    float _last_notch_bandwidth_hz[INS_MAX_INSTANCES];
     float _last_notch_attenuation_dB[INS_MAX_INSTANCES];
 
     // support for updating dynamic filter at runtime
-    uint16_t _last_harmonic_notch_center_freq_hz[INS_MAX_INSTANCES];
-    uint16_t _last_harmonic_notch_bandwidth_hz[INS_MAX_INSTANCES];
+    float _last_harmonic_notch_center_freq_hz[INS_MAX_INSTANCES];
+    float _last_harmonic_notch_bandwidth_hz[INS_MAX_INSTANCES];
     float _last_harmonic_notch_attenuation_dB[INS_MAX_INSTANCES];
     
     void set_gyro_orientation(uint8_t instance, enum Rotation rotation) {
